requirements.txt
numpy
keras
opencv-python
mediapipe
tensorflow
scikit-learn
matplotlib
fastapi
uvicorn
pillow